Add basic Vector and Node structures
For implementation, I need basic structures on which to base the algorithm.
The basic ones are Node and vector.

A vector is a 2 or 3 dimensional structure representing a point in space.

Node is a structure used directly in the algorithm. It allows to calculate
the cost/distance between other nodes. It is based on a 2, or 3-dimensional
vector.