Add configuration of external libraries and application frontend
The change adds external SFML and ImGui libraries. The first version of the GUI
is already prepared and the user can view the grid and interact with it.

CPP and Assembler now link as independent libraries to the app project.
Also added any post-build commands to copy their new changes to the
correct folders.